[Filesystem] create FilesystemTestCase from FilesystemTest
